# docs/OPENAPI.yaml
openapi: 3.0.3
info:
  title: CogniMesh API
  description: |
    Multi-model AI orchestration platform API.
    
    CogniMesh routes work across GPT, Claude, and Kimi through flat-rate subscriptions.
    No API keys. No metered billing. No surprises.
    
    ## Features
    
    - **Multi-Model Routing**: Intelligent routing across 7 AI models
    - **Zero API Billing**: Uses existing subscription clients
    - **5 Operator Surfaces**: CLI, Desktop, VS Code, Copilot, Cursor
    - **58 MCP Tools**: Task management, roadmaps, Claude integration
    - **Company Management**: Multi-tenant organization support
    - **Billing & Budgets**: Cost tracking and budget management
    - **GitHub Integration**: Repository, issue, and PR management
    - **Webhooks**: Event subscriptions with retry logic
    - **Routines**: Scheduled task automation
    - **Approvals**: Workflow approval system
    - **Plugins**: Extensible plugin architecture
    
    ## Authentication
    
    Most endpoints require authentication via:
    - **JWT Token**: `Authorization: Bearer <token>` header
    - **API Key**: `X-API-Key: <key>` header (for service-to-service)
  
  version: 5.0.0
  contact:
    name: CogniMesh Support
    url: https://github.com/LastEld/Ckamal/issues
  license:
    name: MIT
    url: https://github.com/LastEld/Ckamal/blob/main/LICENSE
servers:
  - url: http://localhost:3000/api/v1
    description: Local development server
  - url: https://api.cognimesh.io/api/v1
...
paths:
...
  /api/auth/register:
    post:
      tags:
        - Authentication
      summary: Register new user
      description: Create a new user account
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate and receive access tokens
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /api/auth/logout:
    post:
      tags:
        - Authentication
      summary: User logout
      description: Invalidate session and tokens
      operationId: logout
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Logout successful
  /api/auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Get new access token using refresh token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
  /api/auth/forgot-password:
    post:
      tags:
        - Authentication
      summary: Request password reset
      description: Send password reset email
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Reset email sent
  /api/auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user
      description: Get current user profile
      operationId: getMe
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
    
    put:
      tags:
        - Authentication
      summary: Update current user
      description: Update user profile
      operationId: updateMe
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Profile updated
  /api/auth/api-keys:
    get:
      tags:
        - Authentication
      summary: List API keys
      description: List API keys for current user
      operationId: listApiKeys
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: API keys retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyList'
    
    post:
      tags:
        - Authentication
      summary: Create API key
      description: Create new API key
      operationId: createApiKey
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApiKeyRequest'
      responses:
        '201':
          description: API key created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyResponse'
  /api/auth/api-keys/{id}:
    delete:
      tags:
        - Authentication
      summary: Revoke API key
      description: Revoke an API key
      operationId: revokeApiKey
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: API key revoked

  # Company Endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - name
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 12
        name:
          type: string
          minLength: 1
          maxLength: 100
        companyId:
          type: string
          format: uuid
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
            tokens:
              $ref: '#/components/schemas/TokenResponse'
    TokenResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        role:
          type: string
        createdAt:
          type: string
          format: date-time
    UserProfile:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
            auth:
              type: object
              properties:
                actorType:
                  type: string
                companyId:
                  type: string
                role:
                  type: string
                permissions:
                  type: array
                  items:
                    type: string
    UpdateProfileRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
    CreateApiKeyRequest:
      type: object
      properties:
        name:
          type: string
        permissions:
          type: array
          items:
            type: string
        expiresIn:
          type: integer
          description: Seconds until expiration
        rateLimit:
          type: integer
    ApiKeyResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            key:
              type: string
            apiKey:
              type: object
              properties:
                id:
                  type: string
                name:
                  type: string
                createdAt:
                  type: string
                  format: date-time
    ApiKeyList:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              name:
                type: string
              createdAt:
                type: string

    # Company Schemas
...
security:
  - bearerAuth: []